Add basic role skeleton and doc it
A basic skeleton and molecule has been added to the project which
should assist folks looking to contribute roles to the effort. The
basic skeleton will initialize the all new roles to test with
molecule and provides access to a simple interface develoeprs can
leverage to add new roles into the project.

A minimal implementation of sphinx docs has been added to the project
in support of the new role skeleton.

Change-Id: I7a0fea87226b73b72b40e831922183a527ce901a
Signed-off-by: Kevin Carter <kecarter@redhat.com>