Add test_deps role and basic molecule job
This change creates a role named test_dep which will be used by all
roles tested with molecule. This role is iteself is also tested with
molecule. To ensure the CI is able to run these jobs as independent
tox environments a zuul job was setup which will execute setup the
ci node and execute the tox environment.

Depends-On: Ib1749721b32b21ba1ab661a64097d79037e416f3
Change-Id: Idc784fa03c15c56d0f516f4db2e9cb5eee463fae
Signed-off-by: Kevin Carter <kecarter@redhat.com>