Add helper playbook to create new roles
A playbook has been added to the root of the project which will assist
developers in adding new roles to the project. The playbook will generate
a new role, create a tox entry, and setup zuul for independent CI.

This new playbook capability has been added to the contributing docs.

Change-Id: I726740e1c95b2f38fe7bae2e9a8f1e35feb5a7bf
Signed-off-by: Kevin Carter <kecarter@redhat.com>