Introduce ansible roles for ceph preparation
Create five ansible roles prefixed with "tripleo-ceph-" to
prepare the undercloud to run ceph-ansible. These roles were
previously embedded as tasks in tripleo-heat-templates.
ceph-ansible assumes its working directory persists and is
configured by the user manually. TripleO integration requires
that these steps be automated and these roles perform that
automation.

The roles may be used to create an ansible working directory
with inventory, a ceph-ansible fetch directory, a system uuid
map for node-specific-overrides, and runs ceph-ansible with
the required options.

Add ceph.yaml to the playbooks directory which can be used for
development of these roles with --skip-tags run_ceph_ansible.

Change-Id: I19cd58275a0b8f8f84776a98008cd9b10ceda908