Import octavia roles and playbooks from tripleo-common
This change imports all of the octavia playbooks and roles from
`tripleo-common/playbooks/`. This change ensures all of the resources
imported are meeting the required lint checks for import and structures
the roles such that they'll automatically be installed in the
`usr/share/ansible/roles/` path making them available to the rest of
the tripleo ecosystem.

Change-Id: Ib4ff59a4c372f95cc7a183e8ef724bb1cbf72fed
Signed-off-by: Kevin Carter <kecarter@redhat.com>