Update the _skeleton_ role for proper file generation
This change updates the _skeleton_ role so that the files it generates are
properly formatted. When a new role is generated the resulting code will
now pass all linting checks by default. Before this change several files,
which were assumed to be needing modification by the developer, would fail
lint checks right after generation due to newline issues at the end of the
file.

A test has been added to ensure the role generation process produces
minimally functional role that is 100% compliant.

Change-Id: I28526b022b8c2b5281d2a536231ebcac3b0a7662
Signed-off-by: Kevin Carter <kecarter@redhat.com>