Namespace the install path
The path where we're installing content currently conflicts with
existing path's from existing tripleo repositories. This creates
a conflict as content is likely to exist in more than one
repository while we migrate things to tripleo-ansible. To ensure
we're not creating conflict all of the tripleo-ansible installed
content will now be moved into a namespaced directory.

Change-Id: Iedbf9d1034ee3396d6cb064c64169c21ebcf3e18
Signed-off-by: Kevin Carter <kecarter@redhat.com>