Improve documentation readability
The role documentation titles were too long, and being truncated
in the site sidebar. This change shortens the title names so that
the its easier to navigate the documentation.

An update has been made to the role-addition playbook so that the
new role titles are generated with the shorter length.

Change-Id: Icee3c318ac0049eda586c2696d732415575eaa40
Signed-off-by: Kevin Carter <kecarter@redhat.com>