Import tripleo-container-tag role
This change imports the tripleo-container-tag role and adds testing using
molecule. Tests within this role will be delegated to the host so that
it can natively control both docker and podman containers.

To test this role the "openstack-tox-with-sudo" job is being used which
will allow the job to locally execute against the target node.

Change-Id: I7b30561137cd78b5a3d297fa26115dbfe4458e06
Signed-off-by: Kevin Carter <kevin@cloudnull.com>