Import the tripleo-docker-rm role
The docker-rm role is a duplicate role from tripleo-common which provides
for the same things as the tripleo-container-rm. To ensure we're
developing a consistent user experience the docker-rm role was created
as a symlink to container-rm.

The container-rm role was updated drive consistency across roles by
changing the default "container_cli". Now all roles with the container_cli
variable will have the same default.

Because this role is a symlink no new zuul tests were added for this role.
The interface will be exercised via a new scenario in the container-rm role.

Change-Id: I33ecf735af5c5b0a9a0a9f0c650fc2a04f0864a2
Signed-off-by: Kevin Carter <kecarter@redhat.com>