Remove ansible from our requirements file
In order to have tripleo-ansible added to global requirements
the "ansible" requirement must be removed[0]. This change
transitions the requirement from an upstream maintained file
to our tox config, which we maintained outside of requirements
and is only ever used in local testing. With this package
removed we'll be able to have tripleo-ansible part of the
global requirements sync.

[0] - https://review.opendev.org/#/c/667413/1/projects.txt

Change-Id: Icf1e636bd1fbb73fd6d06a8e43fe55bcbe482122
Signed-off-by: Kevin Carter <kecarter@redhat.com>