Remove irrelevant files
This change rmeoves files that serve no purpose. Most of these things
were brought over on the initial role import and were kept for
consistency. Now that we can test the roles and are confident in the
setup it is time to clean things up by removing files and options
that are empty or pointless.

Change-Id: I832c003a36d82cf37d128e5934aeb5448112e6f8
Signed-off-by: Kevin Carter <kevin@cloudnull.com>