Tweak the role-addition tox job
The role-addition tox job used to make the assumption it would install
docker and run a molecule test, however, due to the fact we're
transitioning away from tox jobs that can modify the underlying
system, the tox environment has been tweaked so that it no longers
makes the assumption docker will be installed via tox.

The tox job will now check for a valid connection to docker or podman
before attempting to run a molecule test against the pseudo role. In
the Zuul environment docker will be present when the resource is
provisioned via the tox-molecule job parent. In a local test
environments, a valid container runtime will need to be deployed,
which documented in the contributing guide. If the role-addition
tox environment is not able to connect to docker or podman a notice
will be displayed linking to the contributing documentation.

The job definition has been updated to ensure it runs anytime the
tox.ini file is changed. This was done because the tox.ini file
is the only mode of execution for this job.

Change-Id: I277edc265e7cd885af55bbf64bbc5e4fda058950
Signed-off-by: Kevin Carter <kecarter@redhat.com>