Fix docs Zuul jobs
The project uses a variant of build-openstack-tox-docs for building
documentation, the publishing job for this is
publish-openstack-tox-docs. Use this publishing job in post queue
instead of the obsolete publish-openstack-sphinx-docs which fails in
this repo since it does not match the build job.

Also, use the release-notes-jobs-python3 instead of adding individual jobs,
this allows updating in a central place. The template does exactly the
same what the jobs do.

Change-Id: I107009c361b594c2493b09fec4e5bc687f103171