Cover the case when the container engine is not available.
In the case the container engine passed in tripleo_container_cli
isn't installed (for example: if we try to remove a docker container
when the instance has podman installed and docker isn't present) then
the role would fail. This patch includes a guard which in the case
the container engine passed in tripleo_container_cli doesn't exist
then nothing is done.
Besides, a check for the container being run in the docker-rm part
was missing, this check is being added too.

Change-Id: Ib5501321fef1a3921fbb17bcfabfb2f5c7c96c41
Related-Bug: #1836531