Use canonical hostname when setting up network ports
The hostname used to set the binding host for Octavia health manager
needs to match the hostname expected by the neutron L2 agent in order
for port binding to succeed. The canonical hostname from the
tripleo ansible inventory might be a better match than the hostname used
in group vars.

Related-Bug: #1837043
Depends-On: I5d23b02ca46a40b39dc25a8367541b0a6e517982
Change-Id: Ie0cf35d95c7632e904bc7be9fbc856b7ff0d21ca