Choose first node as bootstrap node name/ip
Previously when Heat was choosing the bootstrap node, it always chose
the first node due to the ResourceGroup ordering.

When it was switched to ansible, the last node was chosen instead due to
unpredictable dict sorting from the inventory.

This patch restores the previous behavior and makes sure the first node
in a group is chosen as the bootstrap node for the both the name
and ip if the sorted_bootstrap_node variable is set.

The variable is used so that the tripleo-heat-templates patch and this
patch can merge, otherwise they'd block each other from merging.

See I6b93f5b0747c5a11d24615a3bbb5516f9be81401 for the
tripleo-heat-templates patch.

Change-Id: I3d595ea5b84f940a3b2dbc69798f69fe03529c10