Restore AllNodesExtraMapData
This commit is the same as I890bbef764b6bae7cb71cfedaff2ce17829737fb, which was
previously applied to tripleo-common, but lost in the migration to
tripleo-ansible.

Commit a7661065743086961b8ef93056b810e7d2a49eda removed the
all-nodes-config.j2.yaml Heat template. This template had the support
for specifying hieradata to merge into the all_nodes hieradata via the
AllNodesExtraMapData.

This commit restores the merge logic so that the merged hieradata from
the parameter value and the calculated hieradata is rendered in
all_nodes.json.

Change-Id: I49948f8d74c3846321af4f423a4891a71d0fabdc
Closes-Bug: #1839546