Constrain molecule
This change constrains molecule so we're not pulling in broken or otherwise
incompatible code. Recent changes to upstream molecule have introduced some
instability. To ensure we're always testing with an expected release of
molecule a new upper constraint has been added.

Additionally to ensure we're not creating testing gaps, a modification has
been added to the role addition job definition. This will ensure we're
exercising molecule in the event that any of its requirements change in the
future.

Change-Id: Ib12e9a52ace6d3545757ada6863e58843de99ad9
Signed-off-by: Kevin Carter <kecarter@redhat.com>