Import default rules and sort rules
This change imports some embedded rules which were missed when the role
was created. Additionally there was some logic missing that sorted the
rules by comment strings. To ensure we're creating rules in the same way
with the same expected ordering a new fact has been created to ensure
rules are all sorted prior to being written. This fact will now be used
in a single loop which will ensure all rules are written in their
expected orders.

Change-Id: Ie5d0f51d7efccd112847d3f1edf5fd9cdb1edeed
Signed-off-by: Kevin Carter <kecarter@redhat.com>