Add service management support.
The current role for backup and restore will need service
management to insure that all running services on the nodes
are stopped and restarted in order to capture a consistent state.
This change also includes database dump functionality which will
allow the operator to create a clean backup of the database.

Change-Id: Ic75f40a750880be4ea00ed79de4575cdbbe0f190