Move OVN VIP to all_nodes and consider if OVN is configured for a separate VIP
In the same vein as I7ca94dff4acf0816708110b9fe6f78d19dcc7b4d
(Move redis_vip to all_nodes.j2) we want to have the ovn_dbs_vip moved
to all nodes. Because we ultimately want to revert
I0d9eb663405d1113ea84e3c12651a3f0dbdfc75d (Add OvnDbInternal to EndpointMap
and use it for ovn_db_host) as that forces all connections to OVN to go
through the haproxy vip.

We set ovn_dbs_vip to:
A) the ovn_dbs network VIP if net_vip_map.ovn_dbs is not defined
B) to the separate VIP if net_vip_map.ovn_dbs is defined

The separate OVN VIP THT patch that makes the OVN VIP separate is
in I620e37117c26b5b51bf9e1eda91daeb00fdf0f43.

We choose this approach of setting the ovn_dbs_vip in two different
cases because it makes landing all the patches to have OVN live with
a separate VIP a lot simpler (less dependencies).

Testes as follows:
1) Deployed a vanilla tripleo master deployment with this patch and
observed that the OVN VIP is set to the VIP on the internal_api network:
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml ovn_dbs_vip)
   Attributes: inactive_probe_interval=180000 manage_northd=yes master_ip=172.16.2.168 nb_master_port=6641 sb_master_port=6642
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml internal_api_virtual_ip)
   Attributes: inactive_probe_interval=180000 manage_northd=yes master_ip=172.16.2.168 nb_master_port=6641 sb_master_port=6642

2) Also applied I4e4bf0a91751fb4f9e4c7233242cdc5649c421f8 "Revert Add
OvnDbInternal to EndpointMap and use it for ovn_db_host" and observed
that the deploy completes fine and that the OVN VIP is still pointing to
the to the internal_api one:
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml ovn_dbs_vip)
   Attributes: inactive_probe_interval=180000 manage_northd=yes master_ip=172.16.2.220 nb_master_port=6641 sb_master_port=6642
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml internal_api_virtual_ip)
   Attributes: inactive_probe_interval=180000 manage_northd=yes master_ip=172.16.2.220 nb_master_port=6641 sb_master_port=6642

3) Also applied I620e37117c26b5b51bf9e1eda91daeb00fdf0f43
"OVN DBS separate vip" (tht) and Ic62b0fbc0fee40638811a5cd77a5dc5a4d82acf5
"OVN separate vip" (puppet-tripleo) with this review and observed that the OVN VIP is
not the same as the internal_api one:
[root@overcloud-controller-2 hieradata]# hiera -c /etc/puppet/hiera.yaml ovn_dbs_vip
172.16.2.4
[root@overcloud-controller-2 hieradata]# hiera -c /etc/puppet/hiera.yaml internal_api_virtual_ip
172.16.2.48
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml internal_api_virtual_ip)
[root@overcloud-controller-2 ~]# pcs resource show ovn-dbs-bundle |grep $(hiera -c /etc/puppet/hiera.yaml ovn_dbs_vip)
   Attributes: inactive_probe_interval=180000 manage_northd=yes master_ip=172.16.2.4 nb_master_port=6641 sb_master_port=6642

Related-Bug: #1841811

Change-Id: I1d80587752ffca6c3eb5281aa89ea3d7cf5535ce