Check if docker enabled before stopping all containers.
The set of tasks in tripleo-podman/tasks/tripleo_docker_stop.yml
are being used during the upgrade to ensure that all
services are stopped when upgrading from Rocky to Stein.
However, in Train docker does not have to be installed in
the system and this task would fail.
Adding a check before to stop the services only if docker
is installed.

Change-Id: I05fa46d3b53ece83867a2ccd5c931b0cac1209d4