Enhance docker check verifying its socket existence.
The overcloud nodes have podman emulated from the docker command
therefore the check for 'command -v docker' is not a valid check
anymore, as we'll obtain a false possitive. This can end up in
an uglier situation for the case of [0], as if we don't properly
identify that docker is not in the system we might end up removing
the podman containers we have just upgraded.

This patch checks the existence of the docker socket file to
make sure docker is not installed in the system. For podman
we still rely on 'command -v podman'.

[0] - https://github.com/openstack/tripleo-heat-templates/blob/master/deployment/cinder/cinder-api-container-puppet.yaml#L378-L386

Change-Id: I16f1df08ab049600d6139fe4ccf3a680360a3f21