Add local_address to the default hieradata on all hosts
This change will ensure a local_address is always defined within our
generated hieradata. This change imports existing logic into our config
to ensure that the value of the local_address matches our expected
values. We ensure compatibility with legacy installs by inspecting if a
host has IPv6 enabled on the loopback device. In the event that IPv6 is
enabled, the value of local_address will be set to "localhost" otherwise
it will defined as "127.0.0.1".

Closes-Bug: 1843422
Change-Id: I20e69315bacdded4bc2d5b47e18609f130f8abc5
Signed-off-by: Kevin Carter <kecarter@redhat.com>