Open ipv6-icmp traffic by default
puppet-tripleo had special conditions in the rule manifest
to convert the protocol for 'icmp' to 'ipv6-icmp'. This
made it possible to open 'icmp' for ipv4 and 'ipv6-icmp'
for ipv6 using a single rule defenition.

ceph-ansible does not have the logic to support a single
rule. Since the rule to allow icmp traffic is in defined in
the defaults for the role and not in THT. This change uses
the 'ipversion' property for the existing 'icmp' rule and
likewise adds a default rule for 'ipv6-icmp'.

Change-Id: I8b453f7c13c2015aa208ed1bddcdca246cdca58d
Closes-Bug: #1845175