Pivot ovs-dpdk molecule to ansible tasks
This change moves the molecule test setup from the docker file into
ansible tasks. This is being done so that the containers can run
ovs using systemd within the container which removes the need to run
ovs in the foreground, which was proving to be problematic in the
latest releases of centos 7.

Change-Id: I1d1373eeb720e873830cb47ac68efe3421b8b3e0
Signed-off-by: Kevin Carter <kecarter@redhat.com>