Optional flag file for tripleo-transfer to prevent overwriting
When using tripleo-transfer role, optionally a file will be created at
a path specified by tripleo_tranfer_flag_file variable (if set). The
transfer will be prevented in case the file already exists. This is to
ensure the data transfer happens only once and the data is not
overwritten by accident.

Additionally, the README.md file from the tripleo-transfer role is
removed because tripleo-ansible docs are generated by molecule and
this file is not used there.

Change-Id: I3a08949436dcb634df34622c9aeead552dd38443