Set calling_ansible_environment_variables after preset vars
This patch moves calling_ansible_environment_variables after the preset
vars when constructing the ansible-playbook commands for ceph-ansible.

This makes it so that variables set on the command line take precedence
over what is set by these roles. For example, setting a customer
ANSIBLE_CONFIG on the command line would now have that value passed to
the ansible-playbook commands used for ceph-ansible, when previously it
would have been overridden as {{ playbook_dir }}/ansible.cfg.

Change-Id: I2a91ca1aa91d91f6f525f2fba0b093c7ec216e1d