Remove readme files within the roles
The readme files are unnessesary, and unused. All of the role documentation
is generated as part of the tripleo-ansible docs. Because these files are
generally neglected the documentation they provide is at best missing
information and at worst incorrect. This change removes all of the role
readme files and ensures our role generation process no longer creates
the file.

Change-Id: I030d7f9ef54250cef4019d40366febe899f37bd9
Signed-off-by: Kevin Carter <kecarter@redhat.com>