Install packages and load kernel modules before configuring sysctl.
Otherwise we run into dependency problems where a kernel module
may not be loaded yet. This results in the sysctl options not
existing yet in the virtual file system /proc/sys/.

Also add br_netfilter to the required modules for TripleO.
It is required for bridge-nf-call-* sysctl options to work.

Change-Id: Ia28f2fdef34e739801c51828c99e9e6598dd2efb
Related-Bug: #1843259
Signed-off-by: ekultails <ekultails@gmail.com>