paunch: only read container config JSON which starts by "hashed-"
The container-puppet.py generates the hashed-*.json file, which is a copy of
the *step_n.json with a hash of the generated external config added.
(This acts as a salt to enable restarting the container if config
changes)

We want to filter files starting by hashed- otherwise paunch will start
the other containers which will clash with the hashed ones.
Also make sure the container name doesn't have the "hashed-" prefix.

Change-Id: If0f1c6c308cd58f7baa9a8449fbf685ff10f0e0a