Fix network hostnames in tripleo-hosts-entries
The previous Heat behavior used the jinja2 lower() filter on the network
name to create the hostnames, not network.name_lower. As
network.name_lower can container underscores, this was causing an issue
since the expected hostnames were not present in /etc/hosts.

This patch corrects the issue by switching to just using the lower()
filter.

Change-Id: Ia02ca1263590e2b579f2534e99119d7b1cd4b39a