[TRAIN] Backport tripleo-systemd-wrapper (squash)
This is a combination of 4 commits.

This is the 1st commit message:

Implement tripleo-systemd-wrapper role

This patch adds a new role that will be used to manage side containers
with systemd instead of docker.socket or nsenter.  The main use case here
is Neutron, although this role is designed to work with any service.

This role will create a series of systemd files to monitor a file which
gets mounted into a container. Additionally a wrapper script is
generated which is mounted in the container that will provide the
arguments that should be used to launch new containers.

Blueprint: safe-side-containers
Change-Id: I4821b7ca0260e4dfd1717ba976cef700d160f84f
Co-Authored-By: Dan Prince <dprince@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>
Co-Authored-By: Alex Schultz <aschultz@redhat.com>
(cherry picked from commit 699249f1790dd5646556173bf5331e7e71135ad4)

This is the commit message #2:

Remove --rm=true from sidecar container sync

Neutron uses kill-scripts which remove the container after stopping it.
If the container is launched with docker and --rm=true, the container
will automatically be cleaned up and the $(CLI) rm <container id> in the
kill script with error out because the container can't be found.

Related-Bug: #1858662
Change-Id: I3d7940cb0816adce58e0fa778469dcec95302f67
(cherry picked from commit 17d97f2618e56be606c9307551efdaadda8dff69)

This is the commit message #3:

Fix substitution in kill-script

In the kill-script there is a string "Unknown action ${SIG} for
${$CT_NAME} ${CT_ID}" which results in a "bad substitution" error, as
there is no variable named with what the contents of the CT_NAME
environment variable contains.  Remove the extraneous '$'.

Change-Id: I4c76071083bf5cb4f876d3b78c379822a8bd8db1
Fixes-Bug: #1860155
(cherry picked from commit b45d4c6d219e8e27219bca341acdfd634155d6f6)

This is the commit message #4:

Add handling of signal 15 in kill script

The reason bug #1860155 was triggered was because the kill script did
not have a stanza for handling the signal that was passed in, which is
signal 15.  Since signal 15 is unhandled, keepalived processes will
still stick around.  Add handling for signal 15.

Change-Id: I632a3ef5ec137df10f647335f6354589c2316fd0
Related-bug: #1860155
(cherry picked from commit 06dc258a28784db98e44a3de488c204f01b97613)