Fix ssh_known_hosts hostname entries
Previously the tripleo-ssh-known-hosts ansible role used the list of
network names to create the different possible hostnames in the
ssh_known_hosts file. The network names do not match the actual network
hostnames, like internal_api vs. internalapi which results in ssh host
verification to fail and e.g. live migration to fail.

This changes to use the _hostname inventory information to be used to add
hostname information to the ssh_known_hosts file.

Change-Id: If6e99c65084db35a2a4aa1e76b0b89fb5902794a
Closes-Bug: #1852064