Fix ssh_known_hosts hostname entries
The tripleo-ssh-known-hosts ansible role used the list of network names to
create the different possible hostnames in the ssh_known_hosts file. The
network names do not match the actual network hostnames, like internal_api
vs. internalapi which results in ssh host verification to fail and e.g. live
migration to fail.

Initial approach in a5bcbc8d015b792b3546ecbb139506f248dacfe8 had to be reverted
as it resulted in errors when the _hostname inventory information was missing
from the ansible inventory.

Also syncs to use the same approach to create the hostname like in
https://review.opendev.org/693010 to have entries in /etc/hosts and
ssh_known_hosts created in the same way/source.

Closes-Bug: #1852064

Change-Id: Ie04d0f0cd9474070bffd153fa3dddee9f304a14f