tripleo-hieradata: re-enable "service_names"
"service_names" was a useful hieradata which listed the services enabled
for a specific roles vs "enabled_services" which are for all the
services enabled in the cloud, no matter the role.

This is re-added for backward compatibility.

Change-Id: I75b1112089a66cf5db0a2fd651bb24428cf861fd
Related-Bug: #1855138
(cherry picked from commit 114c3c97f400fb99591484e65ed198cbd7ffe584)