Move localhost_address to net_ip_map hieradata
all_nodes is meant to be for hieradata that is the same across all
nodes, while localhost_address could possibly be node specific. This
change moves it from all_nodes to net_ip_map.

This will allow rendering the all_nodes template once and then just
copying it into place, which will be a significant performance
improvement.

Change-Id: Ie3b5f242bc2afce9d395e29db498b3516f8bcc3c