Update regex for presistent save roles
The regex we were using was too general and was not removing the right
rules. This change will make it so that the task will remove lines that
do NOT have a comment, AND contain either "ironic-inspector" OR "neutron-".

This change should provide parity with the old persistent save rule
function we had in puppet.

Closes-Bug: #1854980
Change-Id: I6a7d82938bfb138aa3de875184af8b38babed84e
Signed-off-by: Kevin Carter <kecarter@redhat.com>