Add recursive get key from dict
The docker_config (and other data) that we have in tripleo is a complex
dict but we need a way to fetch a specific key from a larget structure.
This change adds a filter called recursive_get_key_from_dict which will
traverse all the dictionaries in a given dataset and return a list
containing all the values of the requested key.

Partial-Bug: #1870557
Change-Id: I952df4d4a7700eccdb2d49a7fba271518058771a
(cherry picked from commit c072a6bf9d3ca3e264ac9055af13c2a4385895b1)