tripleo-container-manage: show podman commands that are executed
This patch will display the podman commands that are executed for debug
purpose, when check is enabled.
For each container, it'll show the command and also at the very end of
the playbook run, we display ALL commands that were run.

It's replacing the "paunch debug print-cmd" feature that we had with
paunch.

Note: it's disabling async on podman_container module because async +
check don't work together.

story: 2006732
task: 37310

Change-Id: Ic615dd45d8feec3806d5012685e4860d2361b95d