tripleo-container-manage: allow to override a container config
Sometimes an operator might want to override a config parameter for a
specific container, without having to rewrite the JSON on the host.

It can now be done with the new
tripleo_container_manage_config_overrides parameter.

Example:

tripleo_container_manage_config_override:
  mysql:
    image: myregistry.io/tripleomaster/haproxy:latest

It introduces a new attribute to the singledict() filter and its unit
test.

Change-Id: If65fcf25c9e07174b08e44539b0501615b6c72ab